[FEATURE] Allow filtering processes by a time of creation (Issue #9)

+++ params_info/param_base.h
@@ -49,7 +49,7 @@ bool get_int_param(int argc, char *argv[], const char *param, int &param_i,
     if ((param_i + 1) < argc && !is_param(argv[param_i + 1])) {
         char* mode_num = argv[param_i + 1];
         if (isdigit(mode_num[0])) {
-            out_val = (PARAM_T)atoi(mode_num);
+            out_val = (PARAM_T)atol(mode_num);
         }
         else {
             if (callback) {
